Refactor level to loglevel, and undo deprecations
