+++ src/main/java/com/dynatrace/openkit/api/LogLevel.java
@@ -1,0 +1,38 @@
+/**
+ * Copyright 2018-2019 Dynatrace LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.dynatrace.openkit.api;
+
+public enum LogLevel {
+    DEBUG(0),
+    INFO(10),
+    WARN(20),
+    ERROR(30);
+
+    private final int priority;
+
+    LogLevel(int priority) {
+        this.priority = priority;
+    }
+
+    public int getPriority() {
+        return priority;
+    }
+
+    public boolean hasSameOrGreaterPriorityThan(LogLevel other) {
+        return getPriority() >= other.getPriority();
+    }
+}
