+++ src/main/java/com/dynatrace/openkit/api/Logger.java
@@ -28,7 +28,7 @@ public interface Logger {
      * @param level     the level of the log entry
      * @param message   the message to write to the log
      */
-    void log(Level level, String message);
+    void log(LogLevel level, String message);
 
     /**
      * Log with a given level, message and throwable
@@ -37,52 +37,47 @@ public interface Logger {
      * @param message   the message to write to the log
      * @param throwable an instance of a throwable to be attached to the output.
      */
-    void log(Level level, String message, Throwable throwable);
+    void log(LogLevel level, String message, Throwable throwable);
 
     /**
-     * Log with level 'error'
+     * Log with level 'error'.<br>
+     * This is a convenience method for {@link #log(LogLevel, String)}
      *
-     * @deprecated {@link #log(Level, String)}
      * @param message the message to write to the log
      */
-    @Deprecated
     void error(String message);
 
     /**
-     * Log with level 'error'
+     * Log with level 'error'.<br>
+     * This is a convenience method for {@link #log(LogLevel, String, Throwable)}
      *
-     * @deprecated {@link #log(Level, String)}
      * @param message the message to write to the log
      * @param t       an instance of a throwable to be attached to the output
      */
-    @Deprecated
     void error(String message, Throwable t);
 
     /**
-     * Log with level 'warning'
+     * Log with level 'warning'.<br>
+     * This is a convenience method for {@link #log(LogLevel, String)}
      *
-     * @deprecated {@link #log(Level, String)}
      * @param message the message to write to the log
      */
-    @Deprecated
     void warning(String message);
 
     /**
-     * Log with level 'info'
+     * Log with level 'info'.<br>
+     * This is a convenience method for {@link #log(LogLevel, String)}
      *
-     * @deprecated {@link #log(Level, String)}
      * @param message the message to write to the log
      */
-    @Deprecated
     void info(String message);
 
     /**
-     * Log with level 'debug'
+     * Log with level 'debug'.<br>
+     * This is a convenience method for {@link #log(LogLevel, String)}
      *
-     * @deprecated {@link #log(Level, String)}
      * @param message the message to write to the log
      */
-    @Deprecated
     void debug(String message);
 
     /**
